requirements.txt
# Core
torch>=2.0.0
transformers>=4.30.0
faiss-cpu>=1.7.4
numpy>=1.24.0
Pillow>=9.0.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# LLM
openai>=1.0.0

# Demo UI
gradio>=4.0.0

# Eval & CLI
tqdm>=4.65.0
